Build-less search js component
