+++ site/snippets/cheatsheet.header.php
@@ -8,8 +8,7 @@
   <?= css('assets/css/templates/cheatsheet.css') ?>
   <?= css('@auto') ?>
 
-  <?= js('assets/js/index.js', ['defer' => true]) ?>
-  <?= js(['assets/js/site.js'], ['type' => 'module']) ?>
+  <?= js(['assets/js/index.js'], ['type' => 'module']) ?>
   <?= js('assets/js/templates/cheatsheet.js', ['defer' => true]) ?>
 
 </head>
