--- src/js/utils/selector.js
@@ -1,9 +1,0 @@
-const $ = (selector) => {
-  return document.querySelector(selector);
-};
-
-const $$ = (selector) => {
-  return [].slice.call(document.querySelectorAll(selector));
-};
-
-export { $, $$ };
