+++ www/assets/js/index.js
@@ -1,2 +1,30 @@
-!function(e){var n=window.webpackJsonp;window.webpackJsonp=function(t,o,i){for(var u,a,c=0,s=[];c<t.length;c++)a=t[c],r[a]&&s.push(r[a][0]),r[a]=0;for(u in o)Object.prototype.hasOwnProperty.call(o,u)&&(e[u]=o[u]);for(n&&n(t,o,i);s.length;)s.shift()()};var t={},r={1:0};function o(n){if(t[n])return t[n].exports;var r=t[n]={i:n,l:!1,exports:{}};return e[n].call(r.exports,r,r.exports,o),r.l=!0,r.exports}o.e=function(e){var n=r[e];if(0===n)return new Promise(function(e){e()});if(n)return n[2];var t=new Promise(function(t,o){n=r[e]=[t,o]});n[2]=t;var i=document.getElementsByTagName("head")[0],u=document.createElement("script");u.type="text/javascript",u.charset="utf-8",u.async=!0,u.timeout=12e4,o.nc&&u.setAttribute("nonce",o.nc),u.src=o.p+"js/bundle-"+({0:"search"}[e]||e)+".js?id="+{0:"4975eb79f9183cdae52c"}[e];var a=setTimeout(c,12e4);function c(){u.onerror=u.onload=null,clearTimeout(a);var n=r[e];0!==n&&(n&&n[1](new Error("Loading chunk "+e+" failed.")),r[e]=void 0)}return u.onerror=u.onload=c,i.appendChild(u),t},o.m=e,o.c=t,o.d=function(e,n,t){o.o(e,n)||Object.defineProperty(e,n,{configurable:!1,enumerable:!0,get:t})},o.n=function(e){var n=e&&e.__esModule?function(){return e.default}:function(){return e};return o.d(n,"a",n),n},o.o=function(e,n){return Object.prototype.hasOwnProperty.call(e,n)},o.p="/assets/",o.oe=function(e){throw console.error(e),e},o(o.s=0)}({"+Bt+":function(e,n){},0:function(e,n,t){t("vGYV"),t("SgXG"),t("P9yD"),t("+Bt+"),t("rhwI"),t("K4I0"),t("nFT7"),t("XgN2"),t("gpXz"),t("nHye"),t("m56c"),t("IFaM"),t("fDe5"),t("Md71"),t("Kn5K"),t("6J9z"),t("WgIl"),t("FnJk"),t("6314"),t("esmE"),t("2qlg"),t("w6Uh"),t("PprJ"),t("zlsc"),t("koDb"),t("i3Ly"),t("e0Aw"),t("vedj"),t("izEY"),e.exports=t("vg6f")},"2qlg":function(e,n){},6314:function(e,n){},"6J9z":function(e,n){},"7dJq":function(e,n,t){"use strict";Object.defineProperty(n,"__esModule",{value:!0});var r,o=(r=0,function(e){return new Promise(function(n){clearTimeout(r),r=setTimeout(n(),e)})});n.default=o},FnJk:function(e,n){},IFaM:function(e,n){},K4I0:function(e,n){},Kn5K:function(e,n){},Md71:function(e,n){},P9yD:function(e,n){},PSGk:function(e,n,t){"use strict";"kirbyConfig"in window&&window.kirbyConfig.assetsPath&&(t.p=window.kirbyConfig.assetsPath)},PprJ:function(e,n){},SgXG:function(e,n){},WgIl:function(e,n){},XgN2:function(e,n){},akH5:function(e,n,t){"use strict";var r=function(){return function(e,n){if(Array.isArray(e))return e;if(Symbol.iterator in Object(e))return function(e,n){var t=[],r=!0,o=!1,i=void 0;try{for(var u,a=e[Symbol.iterator]();!(r=(u=a.next()).done)&&(t.push(u.value),!n||t.length!==n);r=!0);}catch(e){o=!0,i=e}finally{try{!r&&a.return&&a.return()}finally{if(o)throw i}}return t}(e,n);throw new TypeError("Invalid attempt to destructure non-iterable instance")}}(),o=u(t("sFUq")),i=u(t("7dJq"));function u(e){return e&&e.__esModule?e:{default:e}}var a='<svg viewBox="0 0 5 11" width="5" height="11" aria-hidden="true"><path d="M2.96153846,5.41538462 L0,9.13846154 L1.35384615,10.1538462 L4.73846154,5.92307692 C4.82307692,5.75384615 4.90769231,5.58461538 4.90769231,5.41538462 C4.90769231,5.24615385 4.82307692,5.07692308 4.73846154,4.90769231 L1.35384615,0.676923077 L0,1.69230769 L2.96153846,5.41538462 Z"/></svg>';(0,o.default)(function(){var e=document.querySelector(".js-menu-search");if(e){var n=document.querySelector(".js-menu-search-input");Promise.all([t.e(0).then(t.bind(null,"l3sP")),t.e(0).then(t.bind(null,"S6+x"))]).then(function(t){var o=r(t,2),u=o[0],c=o[1],s=u("S7OGBIAJTV","d161a2f4cd2d69247c529a3371ad3050").initIndex("getkirby-3"),l="",f=[],d=new c(n,{list:[],filter:function(){return!0},sort:!1,item:function(e){var n=document.createElement("li");return"VIEW_ALL_RESULTS_UXfpeDFlmye9rXkXP5wd"===e.label?(n.classList.add("menu-search-view-all"),n.innerHTML="<strong>View all results</strong>"+a):"SEARCH_ERROR_UXfpeDFlmye9rXkXP5wd"===e.label?(n.classList.add("menu-search-error"),n.innerHTML="<strong>Sorry, an error occured. Please try advanced search instead.</strong>"+a):n.innerHTML="<strong>"+e.label+"</strong> <small>"+e.value+"</small>",n}});n.addEventListener("focus",function(){document.documentElement.classList.add("is-menu-search-open")}),n.addEventListener("blur",function(){document.documentElement.classList.remove("is-menu-search-open")}),n.addEventListener("keypress",function(n){(n.key&&"Enter"===n.key||13===n.keyCode)&&e.submit()}),n.addEventListener("keyup",function(){var e=n.value.trim();return e===l||(""===e?(l="",f=[],d.list=f,d.evaluate(),!0):(l=e,e.length<=2||void(0,i.default)(250).then(function(){var t={hitsPerPage:5},r=n.getAttribute("data-filters");r&&(t.filters=r),s.search(e,t,function(e,n){if(f=[],e)return"console"in window&&console.error("Quicksearch error",e),f.push({label:"SEARCH_ERROR_UXfpeDFlmye9rXkXP5wd",value:0}),d.list=f,d.evaluate(),void d.open();for(var t=0,r=n.hits.length;t<r;t++){var o=n.hits[t];f.push({label:o.title,value:o.objectID})}n.nbHits>0&&n.nbHits>n.hitsPerPage&&f.push({label:"VIEW_ALL_RESULTS_UXfpeDFlmye9rXkXP5wd",value:n.nbHits}),d.list=f,d.evaluate(),d.open()})})))}),n.addEventListener("awesomplete-select",function(n){n.preventDefault(),"VIEW_ALL_RESULTS_UXfpeDFlmye9rXkXP5wd"===n.text.label||"SEARCH_ERROR_UXfpeDFlmye9rXkXP5wd"===n.text.label?e.submit():window.location.href="/"+n.text.value})})}})},e0Aw:function(e,n){},esmE:function(e,n){},fDe5:function(e,n){},gpXz:function(e,n){},i3Ly:function(e,n){},izEY:function(e,n){},koDb:function(e,n){},m56c:function(e,n){},nFT7:function(e,n){},nHye:function(e,n){},rhwI:function(e,n){},sFUq:function(e,n,t){"use strict";Object.defineProperty(n,"__esModule",{value:!0}),n.default=function(){var e=arguments.length>0&&void 0!==arguments[0]?arguments[0]:function(){};return new Promise(function(n){"loading"!==document.readyState?(e(),n()):document.addEventListener("DOMContentLoaded",function(){e(),n()})})}},vGYV:function(e,n,t){"use strict";t("PSGk"),t("akH5")},vedj:function(e,n){},vg6f:function(e,n){},w6Uh:function(e,n){},zlsc:function(e,n){}});
-//# sourceMappingURL=index.js.map
+
+import "./components/affiliates.js"
+import "./components/lazyloading.js"
+import "./components/search.js"
+
+// Code highlighting
+import Code from "./components/code.js";
+new Code();
+
+// Menu
+import Menu from "./components/menu.js"
+const menu = document.querySelector(".menu");
+
+if (menu) {
+  new Menu(menu);
+}
+
+// Guide sidebar
+import Sidebar from "./components/sidebar.js"
+const sidebar = document.querySelector(".js-sidebar");
+
+if (sidebar) {
+  new Sidebar(sidebar);
+}
+
+// Tooltips
+import Tooltips from "./components/tooltips.js"
+new Tooltips();
+
+import "./libraries/focus-visible.js";
